Moving Path to it's own dedicated file
